ledger-bindings: rename signTransfer to signAccountTransaction
