+++ packages/ledger-bindings/src/ledger/constants.ts
@@ -1,0 +1,21 @@
+export const INS_VERIFY_ADDRESS = 0x00;
+export const INS_PUBLIC_KEY = 0x01;
+export const INS_SIMPLE_TRANSFER = 0x02;
+export const INS_SIGN_CREDENTIAL_DEPLOYMENT = 0x04;
+export const INS_EXPORT_PRIVATE_KEY_SEED = 0x05;
+/**
+   const INS_TRANSFER_TO_ENCRYPTED = 0x11;
+   const INS_TRANSFER_TO_PUBLIC = 0x12;
+   const INS_ADD_OR_UPDATE_BAKER = 0x13;
+   const INS_REMOVE_BAKER = 0x14;
+   const INS_UPDATE_BAKER_STAKE = 0x15;
+   const INS_UPDATE_BAKER_RESTAKE_EARNINGS = 0x16
+*/
+export const INS_CONFIGURE_DELEGATION = 0x17;
+export const INS_CONFIGURE_BAKER = 0x18;
+
+export const INS_PUBLIC_INFO_FOR_IP = 0x20;
+
+export const INS_UPDATE_CREDENTIALS = 0x31;
+export const INS_SIMPLE_TRANSFER_WITH_MEMO = 0x32;
+export const INS_REGISTER_DATA = 0x35;
